feat(post,comment): create, read, update, delete comments, pagination comments and posts

+++ app/domain/auth.go
@@ -53,7 +53,7 @@ type AuthUseCase interface {
 
 type UserRepository interface {
 	FindByEmail(ctx context.Context, email string) (*User, error)
-	FindByID(ctx context.Context, id int) (*User, error)
+	FindByID(ctx context.Context, id int64) (*User, error)
 	Create(ctx context.Context, tx Transaction, user *User) error
 }
